Add initial LDP-NR support
Files are saved to `./.storage/{nr/uri/path}` but a `StorageAdapter`
interface is in place to allow future implementations to persist to,
e.g. object stores. There is a significant TODO to test thread safety of
the current file implementation.

When an LDP-NR is created we add an LDP-RS (per [LDP
5.2.3.12](http://www.w3.org/TR/ldp/#h-ldpc-post-createbinlinkmetahdr)
which describes the resource. A `describedBy` Link header is added, and
a `powder:describedby` triple connects the two resources in the LDP-NR's
`#metagraph`.

The inverse relationship isn't added. The best way to do  this would
require a special meaning for `powder:describedby` in all LDP-RS
`#metagraph`, which is under consideration.