Better handling of Rack input
The `#parse_graph` implementation failed with a `Rack::Lint::InputStream`,
this is because `#eof?` is called by some `RDF::Reader`s, but is not in
the `InputStream` spec. This *might* have become a problem for Rack
systems that do not use a `StringIO` or other `IO` compliant object for
the input.