Clean up Rack input stream handling
The parser methods used to try to handle Strings, as well as valid Rack
inputs. This led to some unnecessary complexity, giving rise to errors
like those seen in #56. This refactors to handle only IO-like Rack Input
streams.

Most of the work here is just to reorganize the tests.

There is an odd case where `RDFSource#update` feeds the IO stream before
trying to parse a graph. For now, we aggressively `#rewind` the stream
in `#parse_graph` to this problem. This might be unduely costly, and
there is probably a better fix in the long run.