Simplify member triples handling; add support for arbitrary members
Membership triples are now added to the container for all cases. This
simplifies behavior where the location of the membership triples
depended on whether the membership constant URI was a server resource.

Support for members that are not stored by the server is added; e.g.:

    # indirect container
    <> ldp:isMemberOfRelation skos:inScheme ;
       ldp:membershipResource <http://example.com/myScheme> ;
       ldp:insertedContentRelation foaf:primaryTopic .

    # member of indirect container
    <> foaf:primaryTopic <http://example.com/concept> .

The above now succeeds, adding the following triple to the container:

    <http://ex.com/concept> skos:inScheme <http://ex.com/myScheme> .